+++ src/hooks/useHyperparameters.ts
@@ -8,11 +8,17 @@ export const useCustomHyperparameters = (): {
     storeCustomHyperparameters: (hyperparameters: string[]) => void;
     clearCustomHyperparameters: () => void;
 } => {
+    // The type of the value stored for hyperparameters changed between releases 0.17.0 and 0.18.0.
+    // We used to store a list of objects such as `{key: "epoch"}` when we now store a list of strings
+    // In order to not have to do data migration, we opted to store the hyperparameters in a new localStorage variable.
+    // - previously: in custom_hyperparameters_columns
+    // - now: in custom_hyperparameters_columns_
+    // This is also what we did for the `selected_compute_plans_` in the ComputePlans route
     const {
         items: customHyperparameters,
         setItems: storeCustomHyperparameters,
         clearItems: clearCustomHyperparameters,
-    } = useLocalStorageStringItems('custom_hyperparameters_columns');
+    } = useLocalStorageStringItems('custom_hyperparameters_columns_');
 
     return {
         customHyperparameters,